feat(racing-car): 사용자 인풋 검증을 위한 클래스 생성
